requirements.txt
streamlit>=1.32
pandas>=2.1
numpy>=1.26
plotly>=5.18
python-dotenv>=1.0
anthropic>=0.39
openai>=1.40
stravalib>=1.7
requests>=2.31
pydantic>=2.6
tabulate>=0.9
